Renamed time source - related to issue #140
